Yen's K shortest path algorithm working semi
Currently it is working for most cases, but does not take too much detours, for example start=0 and end=5 case, it does not get the route 0-1-4-3-11-10-9-7-5. Not sure why, think it is the assignment of the large_weight that might throw the algorithm off, or premature convergence by including routes with less vertices although having the large_weight costs included. Ideally you would want to have the larger number of vertices included but at the lower total weigth.