Yen and Dijkstra working
Two algorithms are working and not producing any duplicates. Only issue remaining is that of not find excessively long routes as larger weights still have some trade-off. For example the algorithm will rather choose a path with one large weight than with three, and so misses the routes that could have been taken by three.

Also, the removal of edges can sometimes make the graph unconnected, and therefore run into an infinite loop. So rather keep it as it is.